+++ src/entities/note.cpp
@@ -1326,8 +1326,11 @@ QString Note::toMarkdownHtml(QString notesPath, int maxImageWidth,
     // this is a "has not '\w+:\/\/' in it" regular expression
     // see: http://stackoverflow.com/questions/406230/regular-expression-to-match-line-that-doesnt-contain-a-word
     // TODO: maybe we could do that per QTextBlock to check if it's done in comment block?
+    // Important: The `\n` is needed to not crash under Windows if there is just
+    //            an opening `<` and a lot of other text after it
     QRegularExpressionMatchIterator i =
-            QRegularExpression("<(((?!\\w+:\\/\\/)[^<>])+)>").globalMatch(str);
+            QRegularExpression("<(((?!\\w+:\\/\\/)[^<>\n])+)>").globalMatch(str);
+
     while (i.hasNext()) {
         QRegularExpressionMatch match = i.next();
         QString fileLink = match.captured(1);